Refactor & extract logic for building proof nodes
