+++ crates/state/persistent_trie/src/storage.rs
@@ -1,12 +1,15 @@
 use std::{collections::BTreeMap, sync::Arc};
 
 use alloy_rlp::Decodable;
-use alloy_trie::{proof::ProofRetainer, HashBuilder, Nibbles};
+use alloy_trie::Nibbles;
 use edr_primitives::{keccak256, Bytes, StorageKey, B256, U256};
 use hasher::{Hasher, HasherKeccak};
 use revm_state::EvmStorage;
 
-use crate::{persistent_db::PersistentMemoryDB, query::TrieQuery};
+use crate::{
+    persistent_db::PersistentMemoryDB,
+    query::{build_proof_nodes, TrieQuery},
+};
 
 #[derive(Debug)]
 pub(super) struct StorageTrie {
@@ -50,35 +53,18 @@ impl<'a> StorageTrie {
             .map(|key| Nibbles::unpack(keccak256(key)))
             .collect();
 
-        let mut builder =
-            HashBuilder::default().with_proof_retainer(ProofRetainer::new(keys.clone()));
+        let all_proof_nodes = build_proof_nodes(self.trie_query(), keys.clone());
 
-        let mut storage: Vec<(Nibbles, Vec<u8>)> = self
-            .trie_query()
-            .iter()
-            .map(|(key, value)| (Nibbles::unpack(key), value))
-            .collect();
-        storage.sort_by(|(key1, _), (key2, _)| key1.cmp(key2));
-        for (key, value) in storage {
-            builder.add_leaf(key, &value);
-        }
-
-        let _ = builder.root();
-
-        let mut proofs = Vec::new();
-        let all_proof_nodes = builder.take_proof_nodes();
-
-        for proof_key in keys {
-            // Iterate over all proof nodes and find the matching ones.
-            // The filtered results are guaranteed to be in order.
-            let matching_proof_nodes = all_proof_nodes
-                .matching_nodes_sorted(&proof_key)
-                .into_iter()
-                .map(|(_, node)| node);
-            proofs.push(matching_proof_nodes.collect());
-        }
-
-        proofs
+        keys.iter()
+            .map(|proof_key| {
+                // Map over keys so that the result is guaranteed to be in order.
+                all_proof_nodes
+                    .matching_nodes_sorted(proof_key)
+                    .into_iter()
+                    .map(|(_, node)| node)
+                    .collect()
+            })
+            .collect()
     }
 
     fn trie_query(&'a self) -> TrieQuery {
